Refactor project structure and implement Nix development shell
- Moved all public-facing files to a `public/` directory for better organization.
- Renamed assets and relocated them to the `public/assets` directory.
- Replaced custom PHP implementations for database and routing with environment-aware includes.
- Converted database initialization logic into `php/pdo.php`.
- Introduced `shell.nix` for setting up a Nix development shell with `php85` and Composer.
- Updated Docker configuration to adapt to new folder structure.
- Replaced generic MySQL image with version `9.7`.
- Adjusted `compose.debug.yaml` and Apache settings for the new directory layout.
