+++ compose.debug.yaml
@@ -1,8 +1,7 @@
 services:
   mysql:
     container_name: mysql_dev
-    image: mysql
-    command: --default-authentication-plugin=mysql_native_password
+    image: mysql:9.7
     environment:
       MYSQL_ROOT_PASSWORD: test
       MYSQL_DATABASE: test
@@ -15,6 +14,8 @@ services:
       - "8080:80"
     volumes:
         - ./:/var/www/
+    depends_on:
+      - mysql
     environment:
       MYSQL_HOST: mysql_dev
       MYSQL_DATABASE: test
