+++ docker/000-default.conf
@@ -1,8 +1,8 @@
 <VirtualHost *:80>
   ServerAdmin webmaster@localhost
-  DocumentRoot /var/www
+  DocumentRoot /var/www/public
 
-  <Directory /var/www>
+  <Directory /var/www/public>
     Options Indexes FollowSymLinks
     AllowOverride All
     Require all granted
