+++ public/success.php
@@ -1,5 +1,5 @@
 <?php
-include_once('pdo.php');
+include_once('../php/pdo.php');
 
 ?>
