+++ public/user.php
@@ -1,9 +1,9 @@
 <?php
 
-include_once('pdo.php');
-include_once('php/login.inc.php');
-include_once('php/token.inc.php');
-include_once('php/checklogin.php');
+include_once('../php/pdo.php');
+include_once('../php/login.inc.php');
+include_once('../php/token.inc.php');
+include_once('../php/checklogin.php');
 $mysql = new MySQL();
 
 if (!check()) {
