+++ shell.nix
@@ -1,0 +1,5 @@
+{ pkgs ? import <nixpkgs> {}, ... }:
+
+pkgs.mkShell {
+  packages = with pkgs; [php85 php85Packages.composer];
+}
